refactor(editor): keep DEBOUNCE_TIME native in editor-ui, fold getDebounceTime into useDebounce
Per Alex's refinement on PR #34143: DEBOUNCE_TIME is an app-domain catalog
(autosave, resource-search, websocket-disconnect), so it stays natively in
editor-ui's app/constants/durations rather than crossing into a generic lib.
getDebounceTime (parameterized, sessionStorage-only, no DEBOUNCE_TIME ref)
folds into useDebounce.ts and exports from there; the @n8n/composables/durations
module is removed. @/app/constants/durations re-exports getDebounceTime from
@n8n/composables/useDebounce, so all existing call sites keep working unchanged.
CAT-3686.
