+++ packages/frontend/@n8n/composables/src/useDebounce.test.ts
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/vue';
-import { vi, describe, it, expect } from 'vitest';
+import { vi, describe, it, expect, beforeEach } from 'vitest';
 
-import { useDebounce } from './useDebounce';
+import { getDebounceTime, useDebounce } from './useDebounce';
 
 describe('useDebounce()', () => {
 	const debounceTime = 500;
@@ -70,5 +70,43 @@ describe('useDebounce()', () => {
 		button.click();
 
 		expect(mockAsyncFn).toHaveBeenCalledTimes(1);
+	});
+});
+
+describe('getDebounceTime()', () => {
+	beforeEach(() => {
+		sessionStorage.clear();
+	});
+
+	it('returns original time when no multiplier is set', () => {
+		expect(getDebounceTime(100)).toBe(100);
+		expect(getDebounceTime(1500)).toBe(1500);
+	});
+
+	it('applies multiplier from sessionStorage', () => {
+		sessionStorage.setItem('N8N_DEBOUNCE_MULTIPLIER', '0.5');
+		expect(getDebounceTime(100)).toBe(50);
+		expect(getDebounceTime(1500)).toBe(750);
+	});
+
+	it('returns 0 when multiplier is 0', () => {
+		sessionStorage.setItem('N8N_DEBOUNCE_MULTIPLIER', '0');
+		expect(getDebounceTime(100)).toBe(0);
+		expect(getDebounceTime(1500)).toBe(0);
+	});
+
+	it('handles multiplier greater than 1', () => {
+		sessionStorage.setItem('N8N_DEBOUNCE_MULTIPLIER', '2');
+		expect(getDebounceTime(100)).toBe(200);
+	});
+
+	it('defaults to 1 for invalid multiplier values', () => {
+		sessionStorage.setItem('N8N_DEBOUNCE_MULTIPLIER', 'invalid');
+		expect(getDebounceTime(100)).toBe(100);
+	});
+
+	it('rounds to nearest integer', () => {
+		sessionStorage.setItem('N8N_DEBOUNCE_MULTIPLIER', '0.33');
+		expect(getDebounceTime(100)).toBe(33);
 	});
 });
